Add and use a `Gemfile.lock`
This docker image is meant to be a controlled environment where bots can run with a known set of gems installed. If it's not using `Gemfile.lock`, the versions of those gems (and indeed the list of gems itself, if a gem changes its dependencies) can change. Hence we should use one. This one is taken from the current released version of the image, so deploying this change should end up being a no-op.